Remove duplicate unit tests, dispatch to other file or promote to IT
